+++ CHANGELOG.md
@@ -9,9 +9,10 @@
 * add `GetInstanceStats` and `GetInstanceStatsByDate`
 * add `UpdateInstanceIPAddress`
 * add `GetAccountSettings` and `UpdateAccountSettings`
+* add `CreatePayment` `GetPayment` `ListPayments`
+* add `CreateOAuthClient`, `ListOAuthClients`, `GetOAuthClient`, `UpdateOAuthClient`, `DeleteOAuthClient`
 * switched from `metalinter` to `golangci-lint`
 * switched to `go mod` from `dep`
-* add `CreatePayment` `GetPayment` `ListPayments`
 
 <a name-"v0.7.1"></a>